Now we can change for individual Option Set!
The new update includes an option to modify the tax of a group of options or individual options on the OS sections.
You'll find a dropping list with the OS name and when you select this, you'll have a dropping list with the individual options. The first item on the second list is by default 'Change All': this allows us to change all the taxes from that OS massively. Otherwise you can select individualy the options you want to tax update.
Also there's a little change on the way that the items are selected when you want to modify an item. Now they're like the section dropping list, but you can see to which section they belong. This will help us to choose the exact item we want to update in case we have repeated item's names across the menu (in different sections).